Add files via upload
The meshmerize_line_PID code is to check whether your bot works cohesively with the code. I have used an lsa08 sensor and 3 IR sensors to trace the line.
The other two are the final codes for tracing the lines and then finding the shortest path. I have used Arduino UNO and L298N motor driver.
Concept:
It can exhaustively traverse through all possible routes, making a default left turn at every decision making point. This way if there are no loops in the track, the robot will eventually get to the destination. I have kept track of each turn and used it in final run to trace the shortest path.

Sorry i skipped the comments but most of it is easy part. If you are a seasonal coder you may find it difficult to grasp.... no worries ask me directly.